[Refactor][FFT] Consolidate c2c and lut FFT into single FFTC2COp
The LUT-based FFT (SMEM fusion + pre-computed twiddle factors) is
strictly superior to the naive on-the-fly sin/cos implementation.
Merge them into a single FFTC2COp backed by the LUT kernel.

Changes:
- Remove naive fft_c2c kernel (fft_c2c.py) and FFTC2CKernel
- Rename fft_c2c_lut.py → fft_c2c.py, FFTC2CLUTKernel → FFTC2CKernel
- Remove FFTC2CLUTOp, fold LUT logic into FFTC2COp
- Merge test_fft_lut.py into test_fft.py
- Merge bench_fft_lut.py into bench_fft.py (with extended sizes)
- Update kernel and ops __init__.py exports

Net: -444 lines, one Op, one kernel, one test, one benchmark.
