+++ benchmarks/ops/bench_fft.py
@@ -1,3 +1,4 @@
+import math
 from typing import Optional
 
 import pytest
@@ -12,11 +13,14 @@ from tileops.ops import FFTC2COp
 class FFTBenchmarkFixture(FixtureBase):
     PARAMS = [
         ("n, dtype, tune", [
-            (64, torch.complex64, True),
-            (128, torch.complex64, True),
-            (256, torch.complex64, True),
-            (512, torch.complex64, True),
-            (1024, torch.complex64, True),
+            (4096, torch.complex64, True),
+            (8192, torch.complex64, True),
+            (16384, torch.complex64, True),
+            (32768, torch.complex64, True),
+            (4096, torch.complex128, True),
+            (8192, torch.complex128, True),
+            (16384, torch.complex128, True),
+            (32768, torch.complex128, True),
         ]),
     ]
 
@@ -25,16 +29,11 @@ class FFTBenchmark(BenchmarkBase):
 
     def calculate_flops(self) -> Optional[float]:
         n = self.test.n
-        # Cooley-Tukey FFT: log2(N) stages, each with N/2 butterflies.
-        # One butterfly = 1 complex mul (6 FLOPs) + 2 complex adds/subs (4 FLOPs) = 10 FLOPs.
-        # Total: (N/2) * log2(N) * 10 = 5 * N * log2(N) FLOPs.
-        import math
         return 5.0 * n * math.log2(n)
 
     def calculate_memory(self) -> Optional[float]:
         n = self.test.n
         dtype = self.test.dtype
-        # Read N complex + write N complex.
         return 2 * n * torch.empty(1, dtype=dtype).element_size()
