+++ tileops/kernels/fft/__init__.py
@@ -1,4 +1,3 @@
 from .fft_c2c import FFTC2CKernel
-from .fft_c2c_lut import FFTC2CLUTKernel
 
-__all__ = ["FFTC2CKernel", "FFTC2CLUTKernel"]
+__all__ = ["FFTC2CKernel"]
